+++ src/database/migrations/0107-add-reference-to-permission_grant_verb_t.sql
@@ -1,0 +1,1 @@
+ALTER TYPE permission_grant_verb_t ADD VALUE IF NOT EXISTS 'reference';