+++ src/database/operations/authorization/hasSourcePermission.ts
@@ -1,0 +1,8 @@
+import { generateHasPermissionOperation } from '../generators';
+
+const hasSourcePermission = generateHasPermissionOperation(
+	'authorization.hasSourcePermission',
+	'sourceId',
+);
+
+export { hasSourcePermission };